requirements.txt
openai>=1.3.5
llama-index>=0.10.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-llms-openai>=0.1.0
sentence-transformers>=2.2.2
numpy>=1.24.3
torch>=2.1.2
pydantic>=2.0.0
python-dotenv>=1.0.0
PyPDF2
trulens-eval
trulens-apps-llamaindex>=1.0.0
trulens-providers-openai>=1.0.0
pandas
nest_asyncio